# docs/swagger.yaml
openapi: 3.0.0
info:
  title: TradeVault API
  description: |
    A comprehensive trading platform API that provides:
    
    ## Features
    - **User Authentication**: Secure login/register with JWT tokens
    - **OTP Verification**: Email-based OTP system for account security
    - **Real-time Trading**: WebSocket-based live stock updates
    - **Stock Management**: Buy/sell stocks with portfolio tracking
    - **Biometric Authentication**: Advanced security with biometric keys
    - **OAuth Integration**: Google and Apple sign-in support
    
    ## Authentication
    - **App Token**: Used for main application endpoints (JWT_SECRET)
    - **Socket Token**: Used for WebSocket connections (SOCKET_TOKEN_SECRET)
    
    ## Rate Limiting
    - OTP requests: 1 per minute per email
    - Login attempts: 3 attempts before 30-minute lockout
    - PIN attempts: 3 attempts before 30-minute lockout
    
  version: '1.0.0'
  contact:
    name: TradeVault Support
    email: support@tradevault.com
    url: https://tradevault.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.tradevault.com
    description: Production server
security:
  - BearerAuth: []
  - SocketAuth: []
paths:
...
  /auth/send-otp:
    post:
      tags:
        - Authentication
      summary: Send OTP to email
      description: |
        Sends an OTP to the specified email for various purposes:
        - email verification
        - phone number verification
        - password reset
        - PIN reset
      operationId: sendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpRequest'
            example:
              email: "user@example.com"
              otp_type: "email"
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendOtpResponse'
        '400':
          description: Bad request - invalid email or OTP type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify-otp:
    post:
      tags:
        - Authentication
      summary: Verify OTP
      description: |
        Verifies the OTP sent to email for various operations.
        Returns appropriate tokens or confirmation based on OTP type.
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
            example:
              email: "user@example.com"
              otp: "123456"
              otp_type: "email"
              data: null
      responses:
        '200':
          description: OTP verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyOtpResponse'
        '400':
          description: Bad request - invalid OTP or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh-token:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Refreshes the access token using a valid refresh token.
        Supports both app and socket token refresh.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              type: "app"
              refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '400':
          description: Bad request - invalid refresh type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  # ========================================
  # USER MANAGEMENT ENDPOINTS (Protected)
  # ========================================
  /auth/profile:
    get:
      tags:
        - User Management
      summary: Get user profile
      description: Retrieves the authenticated user's profile information
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


    put:
      tags:
        - User Management
      summary: Update user profile
      description: Updates the authenticated user's profile information
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              name: "John Doe"
              gender: "male"
              date_of_birth: "1990-01-01"
              phone_number: "1234567890"

      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
        '400':
          description: Bad request - invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/verify-pin:
    post:
      tags:
        - User Management
      summary: Verify login PIN
      description: Verifies the user's login PIN and returns socket tokens
      operationId: verifyPin
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyPinRequest'
            example:
              login_pin: "1234"
      responses:
        '200':
          description: PIN verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyPinResponse'
        '400':
          description: Bad request - PIN not set or invalid format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid PIN or account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for application authentication.
        Include in Authorization header: `Bearer <token>`
    SocketAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for WebSocket and trading operations.
        Include in Authorization header: `Bearer <token>`
  schemas:
...
    SendOtpRequest:
      type: object
      required:
        - email
        - otp_type
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        otp_type:
          type: string
          enum: [phone, email, reset_password, reset_pin]
          description: Purpose of the OTP
          example: "email"
    SendOtpResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "OTP Sent to registered email address"
    VerifyOtpRequest:
      type: object
      required:
        - email
        - otp
        - otp_type
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        otp:
          type: string
          description: 6-digit OTP code
          example: "123456"
        otp_type:
          type: string
          enum: [phone, email, reset_password, reset_pin]
          description: Purpose of the OTP
          example: "email"
        data:
          type: string
          nullable: true
          description: Additional data (e.g., phone number, new password, new PIN)
          example: "+1234567890"
    VerifyOtpResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "OTP Verified and operation completed successfully"
        register_token:
          type: string
          nullable: true
          description: Registration token (only for email verification)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    RefreshTokenRequest:
      type: object
      required:
        - type
        - refresh_token
      properties:
        type:
          type: string
          enum: [socket, app]
          description: Type of token to refresh
          example: "app"
        refresh_token:
          type: string
          description: Valid refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    RefreshTokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: New JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          description: New JWT refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."


    # ========================================
    # USER MANAGEMENT SCHEMAS
    # ========================================
    ProfileResponse:
      type: object
      properties:
        userId:
          type: string
          description: Unique user identifier
          example: "507f1f77bcf86cd799439011"
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        phone_exist:
          type: boolean
          description: Whether user has set a phone number
          example: false
        name:
          type: string
          description: User's display name
          example: "John Doe"
        login_pin_exist:
          type: boolean
          description: Whether user has set a login PIN
          example: false
        balance:
          type: string
          description: User's account balance (formatted to 2 decimal places)
          example: "50000.00"
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
          description: User's display name
          example: "John Doe"
        gender:
          type: string
          enum: [male, female, other]
          description: User's gender
          example: "male"
        date_of_birth:
          type: string
          format: date
          description: User's date of birth
          example: "1990-01-01"
    UpdateProfileResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success status
          example: true
        data:
          $ref: '#/components/schemas/User'
...
    VerifyPinRequest:
      type: object
      required:
        - login_pin
      properties:
        login_pin:
          type: string
          pattern: '^[0-9]{4}$'
          description: 4-digit login PIN
          example: "1234"
    VerifyPinResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success status
          example: true
        socket_tokens:
          type: object
          properties:
            socket_access_token:
              type: string
              description: JWT token for WebSocket authentication
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            socket_refresh_token:
              type: string
              description: JWT refresh token for WebSocket
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
    User:
      type: object
      properties:
        _id:
          type: string
          description: Unique user identifier
          example: "507f1f77bcf86cd799439011"
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        name:
          type: string
          description: User's display name
          example: "John Doe"
        phone_number:
          type: string
          nullable: true
          description: User's phone number
          example: "+1234567890"
        date_of_birth:
          type: string
          format: date
          nullable: true
          description: User's date of birth
          example: "1990-01-01"
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
          description: User's gender
          example: "male"
        balance:
          type: number
          description: User's account balance
          example: 50000.00
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2024-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-01T00:00:00.000Z"
...
    ErrorResponse:
      type: object
      properties:
        msg:
          type: string
          description: Error message
          example: "Invalid request data"
...